Fixed chat message tests
